+++ CHANGELOG.md
@@ -13,7 +13,7 @@ without compromising stable operations.
 <!-- start-of-changelog -->
 ## 0.72.0 
 - aspect and slope processes now return radians instead of degrees
-
+- YARN batch jobs: configure `spark.driver.maxResultSize` dynamically based on driver memory (max of 5g and driver memory) instead of hardcoded 5g ([#631](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1631))
   
 ## 0.71.0